InputControl: A class for controlling and validating inputs

InputControl: A class for controlling and validating inputs

InputControl: A class for controlling and validating inputs

Details

This class provides methods to validate the inputs of a function.

- The method first checks if the lengths of designs and modes are equal. If not, it raises an error. - It then checks if modes is NULL, and if so, the method returns without performing any further checks. - For each mode in modes, the method sets self$args$mode to the current mode and calls self$check_mode().

The function performs the following checks: - Whether the `clusters` argument is present. - If `clusters` is a single character string "auto", it defaults to automatic cluster estimation and prints a message. - If `clusters` is not a list or if the list contains non-character and non-numeric types, it throws an error.

Check Time Unit

The function performs the following checks: - Whether the `report` argument is present. - If `report` is not a Boolean value (`TRUE` or `FALSE`), it throws an error.

See also

Super classes

SplineOmics::Level4Functions -> SplineOmics::Level3Functions -> SplineOmics::Level2Functions -> InputControl

Public fields

args

A list of arguments to be validated.

Initialize an InputControl object

Methods

Public methods

InputControl$new()

InputControl$auto_validate()

InputControl$check_data_and_meta()

InputControl$check_datas_and_metas()

InputControl$check_datas_descr()

InputControl$check_top_tables()

InputControl$check_mode()

InputControl$check_modes()

InputControl$check_design_formula()

InputControl$check_designs_and_metas()

InputControl$check_spline_params()

InputControl$check_spline_test_configs()

InputControl$check_limma_top_tables()

InputControl$check_adj_pthresholds()

InputControl$check_clusters()

InputControl$check_time_unit()

InputControl$check_report_dir()

InputControl$check_padjust_method()

InputControl$check_report_info()

InputControl$check_report()

InputControl$clone()

Inherited methods

SplineOmics::Level4Functions$create_error_message()

SplineOmics::Level3Functions$check_batch_column()

SplineOmics::Level3Functions$check_condition_time_consistency()

SplineOmics::Level3Functions$check_time_column_pattern()

SplineOmics::Level2Functions$check_columns()

SplineOmics::Level2Functions$check_colums_spline_test_configs()

SplineOmics::Level2Functions$check_data()

SplineOmics::Level2Functions$check_dataframe()

SplineOmics::Level2Functions$check_max_and_min_dof()

SplineOmics::Level2Functions$check_meta()

SplineOmics::Level2Functions$check_spline_params_generally()

SplineOmics::Level2Functions$check_spline_params_mode_dependent()

SplineOmics::Level2Functions$check_spline_type_column()

SplineOmics::Level2Functions$check_spline_type_params()

Method new()

Usage

InputControl$new(args)

Arguments

args

A list of arguments to be validated.

Returns

A new instance of the InputControl class.

Automatically Validate All Arguments

This method automatically validates all arguments by sequentially calling various validation methods defined within the class. Each validation method checks specific aspects of the input arguments and raises an error if the validation fails.

The following validation methods are called in sequence: - self$check_data_and_meta() - self$check_datas_and_metas() - self$check_datas_descr() - self$check_mode() - self$check_modes() - self$check_design_formula() - self$check_designs_and_metas() - self$check_spline_params() - self$check_spline_test_configs() - self$check_adj_pthresholds() - self$check_clusters() - self$check_time_unit() - self$check_report_dir() - self$check_padjust_method() - self$check_report_info() - self$check_report()

Method auto_validate()

Usage

InputControl$auto_validate()

Returns

NULL. The function is used for its side effects of validating input arguments and raising errors if any validation fails.

Check Data and Meta

Method check_data_and_meta()

This function checks the validity of the data and meta objects, ensuring that data is a matrix with numeric values and that meta is a dataframe containing the specified condition column. Additionally, it verifies that the number of columns in the data matrix matches the number of rows in the meta dataframe.

Usage

InputControl$check_data_and_meta()

Arguments

data

A matrix containing numeric values.

meta

A dataframe containing the metadata, including the 'Time' column and the specified condition column.

condition

A single character string specifying the column name in the meta dataframe to be checked.

meta_batch_column

An optional parameter specifying the column name in the meta dataframe used to remove the batch effect. Default is NA.

data_meta_index

An optional parameter specifying the index of the data/meta pair for error messages. Default is NA.

Returns

Returns TRUE if all checks pass. Stops execution and returns an error message if any check fails.

Check Multiple Data and Meta Pairs

Method check_datas_and_metas()

Iterates over multiple data and meta pairs to validate each pair using the `check_data_and_meta` function.

Usage

InputControl$check_datas_and_metas()

Arguments

datas

A list of matrices containing numeric values.

metas

A list of data frames containing metadata.

condition

A character string specifying the column name in the meta dataframe to be checked.

meta_batch_column

An optional parameter specifying the column name in the meta dataframe used to remove the batch effect. Default is NA.

meta_batch2_column

An optional parameter specifying the column name in the meta dataframe used to remove the second batch effect. Default is NA.

Returns

NULL if any check fails, otherwise returns TRUE.

Check Data Descriptions

Method check_datas_descr()

Validates that the data descriptions are character vectors with each element not exceeding 80 characters in length.

Usage

InputControl$check_datas_descr()

Arguments

datas_descr

A character vector of data descriptions.

Returns

No return value, called for side effects.

Method check_top_tables()

Validates that the top tables are a list of dataframes and checks each dataframe using the `check_dataframe` function.

Usage

InputControl$check_top_tables()

Arguments

top_tables

A list of top tables from limma analysis.

Returns

No return value, called for side effects.

Check Mode

Method check_mode()

Validates that the mode is either 'integrated' or 'isolated', which depends on the design formula used in limma.

Usage

InputControl$check_mode()

Arguments

mode

A character string specifying the mode.

Returns

A message indicating the chosen mode if valid; otherwise, an error is thrown.

Method check_modes()

Usage

InputControl$check_modes()

Returns

NULL. The function is used for its side effects of validating each mode and raising errors if any validation fails.

Check Design Formula

Method check_design_formula()

Validates the design formula ensuring it is a valid character string, contains allowed characters, includes the intercept term 'X', and references columns present in the metadata.

Usage

InputControl$check_design_formula()

Arguments

formula

A character string representing the design formula.

meta

A data frame containing metadata.

meta_index

An optional index for the data/meta pair.

Returns

TRUE if the design formula is valid, otherwise an error is thrown.

Method check_designs_and_metas()

Iterates over multiple design formulas and corresponding metadata to validate each pair using the `check_design_formula` function.

Usage

InputControl$check_designs_and_metas()

Arguments

designs

A vector of character strings representing design formulas.

metas

A list of data frames containing metadata.

meta_indices

A vector of optional indices for the data/meta pairs.

Returns

NULL if any check fails, otherwise returns TRUE.

Check Spline Parameters

Method check_spline_params()

Validates the spline parameters both generally and depending on the specified mode.

Usage

InputControl$check_spline_params()

Arguments

spline_params

A list of spline parameters.

mode

A character string specifying the mode ('integrated' or 'isolated').

meta

A dataframe containing metadata.

condition

A character string specifying the condition.

Returns

No return value, called for side effects.

Check Spline Test Configurations

Method check_spline_test_configs()

This function verifies the spline test configurations and associated metadata within the object's arguments. It performs a series of checks on the configurations, including column verification, spline type validation, and ensuring that the degrees of freedom (dof) are within acceptable ranges.

Usage

InputControl$check_spline_test_configs()

Returns

Returns `NULL` if any required arguments (`spline_test_configs` or `metas`) are missing. Otherwise, it performs a series of validation checks.

Check Limma Top Tables Structure

This function checks if the provided limma top tables data structure is correctly formatted. It ensures that the data structure contains exactly three named elements ('time_effect', 'avrg_diff_conditions', and 'interaction_condition_time') and that each element contains dataframes with the correct columns and data types.

Method check_limma_top_tables()

Usage

InputControl$check_limma_top_tables()

Arguments

self

An object containing the data structure to check.

Returns

This function does not return a value. It stops execution if the data structure does not match the expected format.

Check Adjusted p-Thresholds

Method check_adj_pthresholds()

This function checks the validity of the adjusted p-thresholds vector, ensuring that all elements are numeric, greater than 0, and less than 1. If any of these conditions are not met, the function stops execution and returns an error message indicating the offending elements.

Usage

InputControl$check_adj_pthresholds()

Arguments

adj_pthresholds

A numeric vector of adjusted p-thresholds.

Returns

Returns TRUE if all checks pass. Stops execution and returns an error message if any check fails.

Check Clusters

Method check_clusters()

This function verifies the cluster configurations within the object's arguments. It checks if the clusters argument is present and performs validation on its content. If no clusters are specified, it defaults to automatic cluster estimation.

Usage

InputControl$check_clusters()

Method check_time_unit()

This function checks if the provided time unit is valid. The valid time units are: 's' for seconds, 'm' for minutes, 'h' for hours, and 'd' for days. If the time unit is not one of these, the function stops execution and returns an error message.

Usage

InputControl$check_time_unit()

Arguments

time_unit

A character string specifying the time unit. Valid options are 's' for seconds, 'm' for minutes, 'h' for hours, and 'd' for days.

Returns

Returns TRUE if the time unit is valid. Stops execution and returns an error message if the time unit is invalid.

Check and Create Report Directory

Method check_report_dir()

This function checks if the specified report directory exists and is a valid directory. If the directory does not exist, it attempts to create it. If there are any warnings or errors during directory creation, the function stops execution and returns an error message.

Usage

InputControl$check_report_dir()

Arguments

report_dir

A character string specifying the path to the report directory.

Returns

Returns TRUE if the directory exists or is successfully created. Stops execution and returns an error message if the directory cannot be created or is not valid.

Check p-Adjustment Method

Method check_padjust_method()

This function checks if the provided p-adjustment method is valid. The valid methods are: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", and "none". If the method is not one of these, the function stops execution and returns an error message.

Usage

InputControl$check_padjust_method()

Arguments

padjust_method

A character string specifying the p-adjustment method. Valid options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", and "none".

Returns

Returns TRUE if the p-adjustment method is valid. Stops execution and returns an error message if the method is invalid.

Check Report Information

Method check_report_info()

Validates the report information to ensure it contains all mandatory fields and adheres to the required formats.

Usage

InputControl$check_report_info()

Arguments

report_info

A named list containing report information.

Returns

TRUE if the report information is valid; otherwise, an error is thrown.

Check Report

Method check_report()

This function verifies the `report` argument within the object's arguments. It checks if the `report` argument is present and validates its Boolean value.

Usage

InputControl$check_report()

Method clone()

The objects of this class are cloneable with this method.

Usage

InputControl$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.