Level4Functions: A class for level 3 utility functions

Level4Functions: A class for level 3 utility functions

Level4Functions: A class for level 3 utility functions

Details

This class provides methods for creating error messages and checking batch columns.

See also

Methods

Public methods

Level4Functions$create_error_message()

Level4Functions$clone()

Method create_error_message()

This method creates a formatted error message that includes the index of the data/meta pair if provided. If no index is provided, it returns the message as is.

Usage

Level4Functions$create_error_message(message, data_meta_index = NULL)

Arguments

message

A character string specifying the error message.

data_meta_index

An optional parameter specifying the index of the data/meta pair for the error message. Default is NA.

Returns

Returns a formatted error message string. If an index is provided, the message includes the index; otherwise, it returns the message as is.

Method clone()

The objects of this class are cloneable with this method.

Usage

Level4Functions$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.