Package Name: Your Package Name

A detailed description of what your package does.

Key Functions and Classes

List the key functions and classes provided by your package, along with a brief description of each. - Function/Class 1: Description of Function/Class 1 - Function/Class 2: Description of Function/Class 2 - Function/Class 3: Description of Function/Class 3

Package Options

Are there any package-specific options that users can set? If so, please provide details. - Option 1: Description of Option 1 - Option 2: Description of Option 2

Dependencies

Does your package import functions from other packages? If so, please list them. - Package 1: Description of Dependency 1 - Package 2: Description of Dependency 2

Authors

Provide the names and roles of the authors of the package. - Author 1: Role of Author 1 - Author 2: Role of Author 2

Maintainer

Provide the name and contact information of the maintainer. - Name: Maintainer's Name - Email: Maintainer's Email

License

What license is the package distributed under? - License: Your License

Useful URLs

Are there any useful URLs related to the package (e.g., a GitHub repository, documentation site, etc.)? - URL 1: Description of URL 1 - URL 2: Description of URL 2

Additional Information

Is there any additional information that you would like to include in the package documentation? - Additional Info: Description of Additional Info