Build Cluster Hits Report

Generates an HTML report for clustered hits, including plots and spline parameter details, with a table of contents.

Usage

build_cluster_hits_report(
  header_section,
  plots,
  limma_result_2_and_3_plots,
  plots_sizes,
  level_headers_info,
  spline_params,
  adj_pthresholds,
  adj_pthresh_avrg_diff_conditions,
  adj_pthresh_interaction_condition_time,
  mode,
  report_info,
  output_file_path
)

Arguments

header_section

A character string containing the HTML header section.

plots

A list of ggplot2 plot objects.

limma_result_2_and_3_plots

List containing the list of lists with all the plots for all the pairwise comparisons of the condition in terms of average spline diff and interaction condition time, and another list of lists where the respective names of each plot are stored.

plots_sizes

A list of integers specifying the size of each plot.

level_headers_info

A list of header information for each level.

spline_params

A list of spline parameters.

adj_pthresholds

Float vector with values for any level for adj.p.tresh

adj_pthresh_avrg_diff_conditions

Float

adj_pthresh_interaction_condition_time

Float

mode

A character string specifying the mode ('isolated' or 'integrated').

report_info

A named list containg the report info fields. Here used for the email hotkey functionality.

output_file_path

A character string specifying the path to save the HTML report.

Value

No return value, called for side effects.

See also