Check for Required Elements in the SplineOmics Object

This function checks if the given object contains all the required named elements for a specified function type. If any element is missing, it stops the script and provides an informative error message.

Usage

check_splineomics_elements(splineomics, func_type)

Arguments

splineomics

The object to be checked.

func_type

A string specifying the function type. It can be one of "cluster_hits", "create_limma_report", "run_limma_splines", or "explore_data"

Value

None. Stops execution if any required element is missing.