cluster_hits.R contains the exported package function cluster_hits and all the functions that make up the functionality of cluster_hits. cluster_hits clusters the hits of a time series omics datasets (the features that were significantly changed over the time course) with hierarchical clustering of the spline shape. Cluster Hits from Top Tables

Performs clustering on hits from top tables generated by differential expression analysis. This function filters hits based on adjusted p-value thresholds, extracts spline coefficients for significant features, normalizes these coefficients, and applies hierarchical clustering. The results, including clustering assignments and normalized spline curves, are saved in a specified directory and compiled into an HTML report.

cluster_hits(
  splineomics,
  genes,
  adj_pthresholds = c(0.05),
  clusters = c("auto"),
  plot_info = list(y_axis_label = "Value", time_unit = "min", treatment_labels = NA,
    treatment_timepoints = NA),
  report_dir = here::here(),
  analysis_type = "time_effect",
  report = TRUE
)

Arguments

splineomics

An S3 object of class `SplineOmics` that contains all the necessary data and parameters for the analysis, including:

data: The original expression dataset used for differential expression analysis.

meta: A dataframe containing metadata corresponding to the data, must include a 'Time' column and any columns specified by conditions.

design: A character of length 1 representing the limma design formula.

condition: Character of length 1 specifying the column name in meta used to define groups for analysis.

spline_params: A list of spline parameters for the analysis.

meta_batch_column: A character string specifying the column name in the metadata used for batch effect removal.

meta_batch2_column: A character string specifying the second column name in the metadata used for batch effect removal.

limma_splines_result: A list of data frames, each representing a top table from differential expression analysis, containing at least 'adj.P.Val' and expression data columns.

genes

A character vector containing the gene names of the features to be analyzed.

adj_pthresholds

Numeric vector of p-value thresholds for filtering hits in each top table.

clusters

Character or integer vector specifying the number of clusters or 'auto' for automatic estimation.

plot_info

List containing the elements y_axis_label (string), time_unit (string), treatment_labels (character vector), treatment_timepoints (integer vector). All can also be NA. This list is used to add this info to the spline plots. time_unit is used to label the x-axis, and treatment_labels and -timepoints are used to create vertical dashed lines, indicating the positions of the treatments (such as feeding, temperature shift, etc.).

report_dir

Character string specifying the directory path where the HTML report and any other output files should be saved.

analysis_type

String specifying type of limma results (which of the three categories: time_effect, avrg_diff_conditions, and interaction_condition_time ( see limma_result_categories.pdf in inst/descriptions).

report

Boolean TRUE or FALSE value specifing if a report should be generated.

Value

A list where each element corresponds to a group factor and contains the clustering results, including `clustered_hits` data frame, hierarchical clustering object `hc`, `curve_values` data frame with normalized spline curves, and `top_table` with cluster assignments.

See also