Generate Composite Spline Plots

Creates composite spline plots for significant and non-significant features across multiple levels within a condition. One half of one condition comparison HTML (composite spline plots for one 'condition' inside one condition comparison)

Usage

gen_composite_spline_plots(
  internal_combos,
  datas,
  metas,
  spline_test_configs,
  time_unit_label
)

Arguments

internal_combos

A list containing combinations of top tables.

datas

A list of matrices.

metas

A list of metadata corresponding to the data matrices.

spline_test_configs

A configuration object for spline tests.

time_unit_label

A character string specifying the time unit label for plots.

Value

A list containing the composite spline plots and their lengths.

See also