utils scripts contains shared functions that are used by at least two package functions of the SplineOmics package. The level separation is only valid internally in this script, and has no connection to the script level of the respective exported functions scripts. Generate Report HTML

Generates an HTML report with the provided plots, spline parameters, and report information.

generate_report_html(
  plots,
  plots_sizes,
  report_info,
  data = NULL,
  meta = NA,
  topTables = NA,
  enrichr_format = NA,
  level_headers_info = NA,
  spline_params = NA,
  adj_pthresholds = NA,
  report_type = "explore_data",
  feature_name_columns = NA,
  analysis_type = NA,
  mode = NA,
  filename = "report",
  timestamp = format(Sys.time(), "%d_%m_%Y-%H_%M_%S"),
  report_dir = here::here()
)

Arguments

plots

A list of ggplot2 plot objects.

plots_sizes

A list of integers specifying the size of each plot.

report_info

A named list containing report information.

data

A dataframe or a list of dataframes, containing data that should be directly embedded in the HTML report for downloading.

meta

A dataframe, containing metadata that should be directly embedded in the HTML report for downloading.

topTables

List of limma topTables

enrichr_format

List, containing two lists: The gene list and the list of background genes.

level_headers_info

A list of header information for each level.

spline_params

A list of spline parameters, such as dof and type.

adj_pthresholds

Numeric vector with the values for the adj.p.tresholds for each level.

report_type

A character string specifying the report type ('screen_limma_hyperparams' or 'cluster_hits').

feature_name_columns

Character vector with the column names of the annotation information, such as the columns containing the gene names. These column names are used to put the info in the HTML reports on how the descriptions above the individual spline plots where created. This is because those descriptions can be made up of several column values, and the specific columns are then stated in the HTML report on top (e.g gene_uniprotID).

analysis_type

One of the strings "time_effect", "avrg_diff_conditions" , or "interaction_condition_time". Those represent the three different outputs of a limma analysis. For more info on those 3 "categories", see package dir inst/ descriptions/limma_result_categories.pdf.

mode

A character string specifying the mode ('isolated' or 'integrated').

filename

A character string specifying the filename for the report.

timestamp

A timestamp to include in the report filename.

report_dir

A character string specifying the report directory.

Value

No return value, called for side effects.

See also