Calculate Curve Values Based on Top Table Filter

This function filters entries from a given top table based on an adjusted p-value threshold, performs spline interpolation using specified degrees of freedom, and calculates curve values for the selected entries at predefined time points. The function is internal and not exported.

Usage

get_curve_values(top_table, level, meta, condition, spline_params, mode)

Arguments

top_table

A data frame containing data with a column for adjusted p-values and expression averages which indicate the number of degrees of freedom.

level

The specific level of the condition to filter on in the metadata.

meta

Metadata containing time points and conditions.

condition

The name of the condition column in the metadata to filter on.

spline_params

A list of spline parameters for the analysis.

mode

A character string specifying the mode ('isolated' or 'integrated').

Value

A list containing two elements: `curve_values`, a data frame of curve values for each filtered entry, and `smooth_timepoints`, the time points at which curves were evaluated.