Get Header Section

Generates the HTML header section for a report, including the title, header text, and logo. This section also includes the styling for the table and other HTML elements.

get_header_section(title, header_text, report_type, feature_names_formula)

Arguments

title

A string specifying the title of the HTML document.

header_text

A string specifying the text to be displayed in the header of the report.

report_type

A character specifying the type of HTML report.

feature_names_formula

String describing which columns of the annotation info, such as gene and uniprotID, where used to construct the description above the individual spline plots. This is placed in the beginning of the output HTML reports.

Value

A string containing the HTML header section.

Details

The function checks the `DEVTOOLS_LOAD` environment variable to determine the path to the logo image. The logo image is then converted to a base64 data URI and included in the HTML. The header section includes styles for tables, table cells, and header elements to ensure proper formatting and alignment.