Hierarchical Clustering of Curve Values

Performs hierarchical clustering on given curve values. The function adjusts the provided top_table with cluster assignments.

Usage

hierarchical_clustering(curve_values, k, smooth_timepoints, top_table)

Arguments

curve_values

A matrix or data frame of curve values to cluster.

k

The number of clusters to use.

smooth_timepoints

Numeric vector of time points corresponding to columns in curve_values.

top_table

Data frame to be updated with cluster assignments.

Value

A list containing clustering results and the modified top_table.