Plot Composite Splines

Generates composite spline plots for significant and non-significant features based on the specified indices.

Usage

plot_composite_splines(
  data,
  meta,
  spline_test_configs,
  top_table,
  top_table_name,
  indices,
  type,
  time_unit_label
)

Arguments

data

A matrix of data values.

meta

A dataframe containing metadata.

spline_test_configs

A configuration object for spline tests.

top_table

A dataframe containing the top table results.

top_table_name

A character string specifying the name of the top table.

indices

A vector of indices specifying which features to plot.

type

A character string specifying the type of features ('significant' or 'not_significant').

time_unit_label

A string shown in the plots as the unit for the time, such as min or hours.

Value

A list containing the composite plot and its length if plots are generated, FALSE otherwise.

See also