Plot limma Combination Results

Generates plots for pairwise comparisons of hyperparameter combinations using limma results.

Usage

plot_limma_combos_results(
  top_tables_combos,
  datas,
  metas,
  condition,
  spline_test_configs,
  meta_batch_column,
  meta_batch2_column,
  time_unit = time_unit
)

Arguments

top_tables_combos

A list of top tables for each combination.

datas

A list of matrices.

metas

A list of metadata corresponding to the data matrices.

condition

Meta column name that contains the levels.

spline_test_configs

A configuration object for spline tests.

meta_batch_column

A character string specifying the meta batch column.

meta_batch2_column

A character string specifying the second meta batch column.

time_unit

A single character, such as s, m, h, or d, specifying the time_unit that should be used for the plots (s = seconds, m = minutes, h = hours, d = days). This single character will be converted to a string that is a little bit more verbose, such as sec in square brackets for s.

Value

A list of results including hit comparison plots and composite spline plots for each pair of combinations.