Process Combination

Processes a single combination of data, design, spline configuration, and p-threshold to generate LIMMA spline results.

Usage

process_combo(
  data_index,
  design_index,
  spline_config_index,
  pthreshold,
  datas,
  rna_seq_datas,
  metas,
  designs,
  modes,
  condition,
  spline_test_configs,
  feature_names,
  padjust_method,
  ...
)

Arguments

data_index

Index of the data in the datas list.

design_index

Index of the design in the designs list.

spline_config_index

Index of the spline configuration in the spline_test_configs list.

pthreshold

The p-value threshold for significance.

datas

A list of data matrices

rna_seq_datas

A list of RNA-seq data objects, such as the voom object derived from the limma::voom function.

metas

A list of metadata corresponding to the data matrices.

designs

A list of design matrices.

modes

A character vector containing 'isolated' or 'integrated'.

condition

A single character string specifying the condition.

spline_test_configs

A configuration object for spline tests.

feature_names

A character vector of feature names.

padjust_method

A single character string specifying the p-adjustment method.

...

Additional arguments.

Value

A list of top tables from the LIMMA spline analysis.

See also