Process Enrichment Results

Process enrichment results for visualization.

Usage

process_enrichment_results(
  all_db_results,
  enrichment_results,
  adjP_threshold,
  column_name,
  count_column_name,
  background = FALSE
)

Arguments

all_db_results

A list of data frames containing enrichment results for all databases.

enrichment_results

A list of data frames containing enrichment results for individual databases.

adjP_threshold

The threshold for adjusted p-values.

column_name

The name of the column to store adjusted p-values.

count_column_name

The name of the column to store gene counts.

background

Logical indicating whether background ratios are included.

Value

A list of data frames containing processed enrichment results.