Process and Encode Data Field for Report

This function processes a given field, encodes the associated data as base64, and generates a download link for the report. It handles different types of fields including data, meta, top tables, and Enrichr formatted gene lists.

Usage

process_field(
  field,
  data,
  meta,
  topTables,
  report_info,
  encode_df_to_base64,
  report_type,
  enrichr_format
)

Arguments

field

A string specifying the field to process.

data

A dataframe containing the main data.

meta

A dataframe containing meta information.

topTables

A dataframe containing the results of differential expression analysis.

report_info

A list containing additional report information.

encode_df_to_base64

A function to encode a dataframe to base64.

report_type

A string specifying the type of report.

enrichr_format

A list with the formatted gene lists and background gene list.

Value

A string containing the HTML link for downloading the processed field.