run_gsea()

This function generates a Gene Set Enrichment Analysis (GSEA) report based on clustered hit levels, gene data, and specified databases. It processes the input data, manages GSEA levels, and produces an HTML report with plots.

Usage

run_gsea(
  levels_clustered_hits,
  databases,
  report_info,
  clusterProfiler_params = NA,
  plot_titles = NA,
  universe = NULL,
  report_dir = here::here()
)

Arguments

levels_clustered_hits

A list of clustered hits at different levels.

databases

A list of databases for the gene set enrichment analysis.

report_info

A list containing information for the report generation.

clusterProfiler_params

Additional parameters for the GSEA analysis, default is NA. Those include adj_p_value, pAdjustMethod, etc (see clusterProfiler documentation).

plot_titles

Titles for the plots, default is NA.

universe

Enrichment background data, default is NULL.

report_dir

Directory where the report will be saved, default is `here::here()`.

Value

A list of plots generated for the GSEA report.