screen_limma_hyperparams.R contains the exported package function screen_limma_hyperparams and all the functions that make up the functionality of screen_limma_hyperparams. screen_limma_hyperparams runs the other package function, run_limma_splines, for a time series omics dataset, for different hyperparameters. Such are for example degree of freedom of the spline, type of spline, limma design formula, and different versions of the data (full data vs. outliers removed). This can result in several combinations, and it is tedious analyzing the combination in an organised and structured manner. Therefore, this function streamlines that process. Limma Hyperparameters Screening

This function screens through various combinations of hyperparameters for limma analysis, including designs, modes, and degrees of freedom. It validates inputs, generates results for all combinations, and plots the outcomes. Finally, it may also be involved in generating an HTML report as part of a larger analysis workflow.

screen_limma_hyperparams(
  datas,
  datas_descr,
  metas,
  designs,
  condition,
  spline_test_configs,
  report_info,
  report_dir = here::here(),
  adj_pthresholds = c(0.05),
  meta_batch_column = NA,
  meta_batch2_column = NA,
  time_unit = "min",
  padjust_method = "BH"
)

Arguments

datas

A list of data frames containing the datasets to be analyzed.

datas_descr

A description object for the data.

metas

A list of data frames containing metadata for each dataset in `datas`.

designs

A character vector of design formulas for the limma analysis.

condition

A single character string specifying the condition.

spline_test_configs

A configuration object for spline tests.

report_info

An object containing report information.

report_dir

A non-empty string specifying the report directory.

adj_pthresholds

A numeric vector of p-value thresholds for significance determination.

meta_batch_column

A character string specifying the meta batch column.

meta_batch2_column

A character string specifying the second meta batch column (the limma function removeBatchEffect supports a maximum of two batch columns.)

time_unit

A character string specifying the time unit label for plots.

padjust_method

A character string specifying the method for p-value adjustment.

Value

Returns a list of plots generated from the limma analysis results. Each element in the list corresponds to a different combination of hyperparameters.