Create and update the SplineOmics object =====

Creates a SplineOmics object containing variables that are commonly used across multiple functions in the package.

Usage

create_splineomics(
  data,
  meta,
  condition,
  annotation = NULL,
  report_info = NULL,
  meta_batch_column = NULL,
  meta_batch2_column = NULL,
  feature_name_columns = NULL,
  design = NULL,
  spline_params = NULL,
  preprocess_rna_seq = FALSE,
  normalization_fun = NULL,
  padjust_method = "BH"
)

Arguments

data

The actual omics data.

meta

Metadata associated with the omics data.

condition

A condition variable.

annotation

A dataframe with the feature descriptions of data (optional).

report_info

A list containing report information such as omics data type, data description, data collection date, analyst name, contact info, and project name (optional).

meta_batch_column

Column for meta batch information (optional).

meta_batch2_column

Column for secondary meta batch information (optional).

feature_name_columns

Character vector containing the column names of the annotation info that describe the features. This argument is used to specify in the HTML report how exactly the feature names displayed above each individual spline plot have been created. Use the same vector that was used to create the row headers for the data matrix!

design

A design matrix or similar object (optional).

spline_params

Parameters for spline functions (optional).

preprocess_rna_seq

Boolean specifying whether to preprocess RNA seq

normalization_fun

Function used for normalizing RNA-seq. Must take as input the y of: y <- edgeR::DGEList(counts = raw_counts) and output the y with the normalized counts.

padjust_method

Method for p-value adjustment, one of "none", "BH", "BY", "holm", "bonferroni", "hochberg", or "hommel". Defaults to "BH" (Benjamini-Hochberg).

Value

A SplineOmics object.

Details

Description ———– Contains the functions to create and update a SplineOmics object. This object is used to collect function arguments, that are equivalent for more than one exported function of the SplineOmics package. Additionally

Functions ——— - create_splineomics: Create a SplineOmics object - update_splineomics: Add additional arguments to the SplineOmics object or overwrite existing arguments.

Classes ——- None

Notes —– None Create a SplineOmics object