The exported function download_enrichr_databases allows to selects & download Enrichr databases for local use with clusterProfiler. Download Enrichr Databases

This function downloads gene sets from specified Enrichr databases and saves them to a specified output directory as a .tsv file. The file is named with a timestamp to ensure uniqueness.

download_enrichr_databases(gene_set_lib, output_dir = here::here())

Arguments

gene_set_lib

A character vector of database names to download from Enrichr.

output_dir

A character string specifying the output directory where the .tsv file will be saved. Defaults to the current working directory.

Value

This function does not return a value but saves a .tsv file in the specified directory containing the gene sets from the specified Enrichr databases.