run_limma_splines.R contains the exported package function run_limma_splines and all the functions that make up the functionality of run_limma_splines. run_limma_splines performs a limma analysis, using splines, to assign a p-value to every feature of a time series omics dataset, to find out which features are significantly changed over the time course. Run Limma Analysis with Spline Interpolation for Hyperparameter Screening

This function conducts differential expression analysis using the Limma package, incorporating spline interpolation to model the effect of various experimental factors across different levels. It supports both isolated and integrated modes for within-level analysis and between-level comparison, adjusting for multiple degrees of freedom corresponding to the factors under investigation.

run_limma_splines(splineomics)

Arguments

splineomics

A SplineOmics object, containing data, meta, design, condition, and spline_params.

Value

The SplineOmics object, updated with a list with three elements: - `time_effect`: A list of top tables for each level with the time effect. - `avrg_diff_conditions`: A list of top tables for each comparison between the levels. The comparison is the average difference of the values. - `interaction_condition_time`: A list of top tables for each comparison between levels. The comparison is the interaction between the condition and the time.